+++ src/carnot/exec/filter_node.h
@@ -26,7 +26,7 @@ class FilterNode : public ProcessingNode {
   Status OpenImpl(ExecState* exec_state) override;
   Status CloseImpl(ExecState* exec_state) override;
   Status ConsumeNextImpl(ExecState* exec_state, const table_store::schema::RowBatch& rb,
-                         int64_t parent_index) override;
+                         size_t parent_index) override;
 
  private:
   std::unique_ptr<VectorNativeScalarExpressionEvaluator> evaluator_;
